+++ CHANGELOG.rst
@@ -3,6 +3,21 @@ Changelog
 ================================
 
 
+Version 0.1.1 (16-03-2019)
+===========================================================
+
+Added
+````````````````````````````````
+
+* Added command to cleanup jobs with active status to an inactive status,
+  should be used when restarting rippy. Enabled on Docker container too.
+
+Changed
+````````````````````````````````
+
+* Failing avgle if captcha is gone but job isn't progressing.
+
+
 Version 0.1.0 (15-03-2019)
 ===========================================================